feat: add ViewProfileActionUri parsing
